feat(plugin): adapter plugin -- snowflake and sqlserver
