+++ configure/etc/conf/plugins.properties
@@ -34,6 +34,8 @@ datacap-plugin-oracle=plugin/datacap-plugin-oracle/pom.xml
 datacap-plugin-phoenix=plugin/datacap-plugin-phoenix/pom.xml
 datacap-plugin-postgresql=plugin/datacap-plugin-postgresql/pom.xml
 datacap-plugin-presto=plugin/datacap-plugin-presto/pom.xml
+datacap-plugin-snowflake=plugin/datacap-plugin-snowflake/pom.xml
+datacap-plugin-sqlserver=plugin/datacap-plugin-sqlserver/pom.xml
 datacap-convert-txt=convert/datacap-convert-txt/pom.xml
 datacap-convert-json=convert/datacap-convert-json/pom.xml
 datacap-convert-xml=convert/datacap-convert-xml/pom.xml
