--- plugin/datacap-jdbc-snowflake/src/main/java/io/edurt/datacap/plugin/jdbc/snowflake/SnowflakeAdapter.java
@@ -1,13 +1,0 @@
-package io.edurt.datacap.plugin.jdbc.snowflake;
-
-import io.edurt.datacap.spi.adapter.JdbcAdapter;
-import io.edurt.datacap.spi.connection.JdbcConnection;
-
-public class SnowflakeAdapter
-        extends JdbcAdapter
-{
-    public SnowflakeAdapter(JdbcConnection jdbcConnection)
-    {
-        super(jdbcConnection);
-    }
-}
