--- plugin/datacap-jdbc-sqlserver/src/main/java/io/edurt/datacap/plugin/jdbc/sqlserver/SqlServerAdapter.java
@@ -1,13 +1,0 @@
-package io.edurt.datacap.plugin.jdbc.sqlserver;
-
-import io.edurt.datacap.spi.adapter.JdbcAdapter;
-import io.edurt.datacap.spi.connection.JdbcConnection;
-
-public class SqlServerAdapter
-        extends JdbcAdapter
-{
-    public SqlServerAdapter(JdbcConnection jdbcConnection)
-    {
-        super(jdbcConnection);
-    }
-}
