--- plugin/datacap-jdbc-sqlserver/src/main/java/io/edurt/datacap/plugin/jdbc/sqlserver/SqlServerVersion.java
@@ -1,20 +1,0 @@
-package io.edurt.datacap.plugin.jdbc.sqlserver;
-
-public enum SqlServerVersion
-{
-    V2017("2017"),
-    V2019("2019"),
-    V2022("2022");
-
-    private String version;
-
-    SqlServerVersion(String version)
-    {
-        this.version = version;
-    }
-
-    public String getVersion()
-    {
-        return this.version;
-    }
-}
