+++ plugin/datacap-plugin-snowflake/src/main/java/io/edurt/datacap/plugin/SnowflakePlugin.java
@@ -1,0 +1,5 @@
+package io.edurt.datacap.plugin;
+
+public class SnowflakePlugin
+        extends Plugin
+{}
