+++ plugin/datacap-plugin-snowflake/src/main/java/io/edurt/datacap/plugin/SnowflakeService.java
@@ -1,0 +1,19 @@
+package io.edurt.datacap.plugin;
+
+import io.edurt.datacap.spi.PluginService;
+
+public class SnowflakeService
+        implements PluginService
+{
+    @Override
+    public String driver()
+    {
+        return "net.snowflake.client.jdbc.SnowflakeDriver";
+    }
+
+    @Override
+    public String connectType()
+    {
+        return "snowflake";
+    }
+}
