+++ plugin/datacap-plugin-snowflake/src/main/resources/META-INF/services/io.edurt.datacap.plugin.Plugin
@@ -1,0 +1,1 @@
+io.edurt.datacap.plugin.SnowflakePlugin
