+++ plugin/datacap-plugin-sqlserver/src/main/java/io/edurt/datacap/plugin/SqlServerPlugin.java
@@ -1,0 +1,5 @@
+package io.edurt.datacap.plugin;
+
+public class SqlServerPlugin
+        extends Plugin
+{}
